Add Cython .c files and break build in 2 parts
Separate the generation of .c files by Cython and building the C
extensions. setup.py now takes a --cython argument that triggers Cython
conversion. If not set, will just build the given .c files.

To avoid problems with readthedocs, modules try to import the Cython
mods and upon failure substitute functions with a mock that raises an
exception.

Docs of gravmag.prism is now mostly in the module docstring because all
functions are in the Cython module. Can still get the docstrings of the
functions but not in the built docs.

TODO: Need to include the .c files in Manifest