Switched to positional and model args in Misfit
Misfit takes 2 dicts, positional and model, that describe the positional
arguments of the data (x, y, z, t, etc) and model parameters (density,
structural index, etc). This unifies subclass attributes which are
accessed using self.positional and self.model.

Implemented methods in Misfit to use only a subset of the data. This
could be used to implement the ExpandingWindow class in gravmag.euler to
substitute the Classic.exapanding_window method.