Benchmark and profile new blocky.Gravity class
Compare the performance of Gravity.planting with the harvester module.
Gravity.planting is a bit slower because of many hash calculations in
Gravity.predicted. Might be able to overcome this by passing optional
arg `fromcache=True` to value and shape_of_anomaly that bypasses the
call to predicted and gets the array straight from cache.