prism kernel funcs call Cython code
There was a big difference in the Cython vs numpy code because of the
dummy variable @birocoles was using. This is what caused the huge
difference in bx, by, bz and why the tests were failing. By removing the
dummy, the differenced dropped to below 10-8. I reduced the accuracy to
10**-8 for the tests of kernel funcs and bx, etc, because it was good
enough (6-8 orders of magnitude lower than the function values).