Tesseroid fwd modeling takes an open Pool
The forward modeling functions in gravmag.tesseroid take an option
multiprocessing.Pool object to run things in parallel. This way, a pool
object can be reused for multiple calls, avoiding the repeated spawning
of processes.

Refactored the function code in the process to have less repetition. The
_dispatcher function now: checks arguments, spawns the pool if
necessary, dispatches the work, and returns the result.