Fixed bug in PointGrid mesh ordering
The grid was mixing up x (North-South, rows) with y (East-West,
columns). This is a legacy bug that was fixed in gridder.regular but
didn't propagate the PointGrid because it wasn't using "regular" to
generate its grid. It now uses (also avoids duplicate code) and fixes
the order. Had to fix the Bk matrix in the PEL equivalent layer as well.