Gambiarra to make PR docs find their CSS files
The PR docs couldn't find their CSS files because sphinx uses relative
paths. This is a hack to make sphinx insert the site url in front of
links to try to overcome this. A problem is that I cant overwrite the
css macro in the template. A work  around is to define the links again
with the url.