Tesseroid should warn if ignoring cells too small
The code ignores input tesseroids that are too small but does so
silently. It should spit out a warning.

Also, when dividing below a size 1e-6 (horizontal) I get a
ZeroDivisionError. The code should not divide such small tesseroids but
warn the user of this.

Made two tests that check this. They fail still because I haven't made
the correction.