Apply tight_out to Moulder app and fix grid on/off
Was calling ax.grid() without arguments so 2 consecutive calls were
toggling the grid on and off when it should just make sure that it's on.
Fixed by calling ax.grid(True) every time.

Apply tight_out by making a draw_event and resize_event callback that
calls self.figure.tight_layout(). Need to store the figure for this so
doing that in _figure_setup