Style fixes for the website before 0.5 release (#320)
Fixes #307 and #285

Fixes some layout problems with the navbar (it was collapsing on smaller screens, 
like notebooks). 
Remove the cookbook link from the navbar because we're favoring the gallery. 
Include a link on the navbar to the API docs directly. 
Remove the github link. 
Some style changes to the homepage.