Fixes error with gravmag.tesseroid on numba 0.28 (#328)
The `fatiando.gravmag.tesseroid` code fails with the latest numba (0.28).  This
was uncaught because installing Mayavi limits numba to 0.26, where the error
doesn't occur. The problem was in the `engine_factory` function. The `looplift`
argument to `numba.jit` doesn't seem to work anymore. The fix was to move the
array allocations out of the jitted function into the calling Python function.
This breaks the numpy only functions. They were too slow to be usable anyway so
I removed them from the code.

Caught the error by running the CI builds without installing Mayavi, but kept a
separate build with Mayavi to generate the documentation HTML. This forced to
configure the CI builds. Now there are separate builds for compiling the docs,
testing in Linux, and testing in OSX.