Remove Mayavi plots from the gallery (#327)
Getting the docs to build with Mayavi plots in the gallery is a dependency
nightmare! Needed to patch traits and pyface but suddenly all of that broke (no
surprise). It's not worth all the headaches involved.

I removed all Mayavi examples from the Gallery and all patches from the
requirements files.