Only build pushes to the master branch (#351)
Right now, when a PR is opened from a branch in the main
fatiando/fatiando repo, the CI systems will start two builds: 1. the PR,
2. the push to the feature branch. This is not necessary and only causes
a longer build time. So I'm setting up the CI system to only build
pushes to the master branch (needed for building the docs in
fatiando.org/dev).