Start defining the API reference as a single page (#353)
Finishes work started in #266 

Creates a single API page that lists all functions/classes in the
library. The list is done for gridder, mesher, utils, and datasets. The
other subpackages still use the list defined in the package docstring
and will need to be updated in later PRs.

The list is defined in the `doc/api.rst` file. When adding a new function
or class to a module, add the corresponding entry to the `api.rst` file.

All links are updated and links to packages in the changelog were
removed (they would break eventually again anyway).