Rewrite the from_image function in datasets (#363)
Replace the `utils.fromimage` function with `datasets.from_image`. The
function is more complete and uses the unique colors to create a
template (instead of mapping grey-scale to a range). Includes tests and
sample images for the gallery and docstring example.

Removed the `utils.fromimage` function and the `SquareMesh.img2prop`
method that relied on it.