Normalize the colorbar in the seismic_image plot (#366)
If vmin and vmax are not specified, will automatically set them so that
the colorbar is centered on zero. This makes the plot nicer when using a
diverging colorbar like RdBu.

The `dt` argument for `seismic_image` and `seismic_wiggle` doesn't have
a default value anymore, as it shouldn't. This is specific to each data
set.