Fix the problem with the PEL test (#375)
* Fix failing tests due to numpy 1.12 update (#373)

A few tests started breaking with the new numpy update.
This PR fixes these tests.

The code for seismic.conv uses np.floor and np.ceil to get 
index number for numpy arrays. 
This now causes errors. 
Wrap both calls in `int(...)` to fix the problem.

* Fix failing pole reduction test in Windows 32bit (#374)

The pole reduction test for the PEL is failing in Windows 32bit
for some reason (probably some floating point rounding problem).
This PR fixes that test by adjusting the amount of regularization.

* Add AppVeyor badge to README (#372)

To inform that we test on Linux and Windows and provides
a quick link to the latest CI status.