Mark PEL pole reduce test as expecting to fail (#380)
This test keeps failing randomly in different CIs and OSs. I'm marking
it as "expected to fail" so that it doesn't jam our CI builds.

This is not a fix.

Will need a better test case that doesn't fail. Maybe weight the misfit
function so that the numbers aren't so tiny.