Use flake8 and pylint instead of pep8 (#382)
flake8 is better than pep8 and easier to configure.

Added a "make lint" command to run pylint static analysis.
This won't break the build because there are so many failures
that it will take a long time to sort through.

Added "make check3" command to use pylint to check Python 3
compatibility. This will break the build if something doesn't pass.
It will make it easier for new code to be ported to Python 3.
I put in the effort of making the necessary adjustments for these
checks to pass.