Implement regression tests for gravmag.polyprism (#395)
Simplify the tests of `polyprism` vs `prism` and add regression tests against saved
results, like was done for `sphere` in #364. 

This allows us to test more complex polygonal prisms (an ellipse for example) and 
catch regressions that might affect both `polyprism` and `prism` at the same time.

Test coverage decreases because the numpy alternative implementation is not
tested anymore. This will become the main version in #368 so it's not a big deal 
right now.